PyDispatch/requirements.txt
customtkinter>=5.2.0
mysql-connector-python>=8.0.0
pillow>=10.0.0
